RL-Model-main/RL-Model-main/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
openai>=1.20.0
requests>=2.31.0
python-multipart>=0.0.9
httpx>=0.27.0
PyYAML>=6.0
